+++ app/components/ClientOnly.tsx
@@ -1,9 +1,7 @@
 "use client";
 
-import React, { useState, useEffect, useLayoutEffect } from "react";
+import React, { useState, useEffect } from "react";
 import Loader from "./Loader";
-import { useTranslation } from "react-i18next";
-import { useAppSelector } from "../context/state/hooks";
 import "@/app/i18n";
 
 interface ClientOnlyProps {
@@ -12,16 +10,10 @@ interface ClientOnlyProps {
 
 const ClientOnly: React.FC<ClientOnlyProps> = ({ children }) => {
   const [hasMounted, setHasMounted] = useState(false);
-  const { i18n } = useTranslation();
-  const bgLocalization = useAppSelector((state) => state.bgLocalization.value);
 
   useEffect(() => {
     setHasMounted(true);
   }, []);
-
-  useLayoutEffect(() => {
-    i18n.changeLanguage(bgLocalization);
-  }, [i18n, bgLocalization]);
 
   if (!hasMounted) return <Loader />;