+++ app/components/pages/main/index/bottom/Left.tsx
@@ -8,7 +8,7 @@ const LeftBottom = () => {
       <div className="flex flex-col w-full">
         <div className="aspect-[3/4] h-full relative overflow-hidden">
           <Image
-            alt="Listing"
+            alt="Mobile app"
             src={"/images/mobileapp.png"}
             className="object-cover h-full w-full"
             fill