+++ app/context/TRQs/favorites/mutations/useRemoveListingFromFavorites.ts
@@ -5,8 +5,10 @@ import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { removeListingFromFavoritesAsync } from "pawpal-fe-common/favorites-api";
 import toast from "react-hot-toast";
 import useFavoriteListings from "../../listings/useFavoriteListings";
+import { useTranslation } from "react-i18next";
 
 const useRemoveListingFromFavorites = () => {
+  const { t } = useTranslation();
   const queryClient = useQueryClient();
   const { handleError } = useGlobalErrorHandler();
 
@@ -15,7 +17,7 @@ const useRemoveListingFromFavorites = () => {
       removeListingFromFavoritesAsync(targetItemId),
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: [useFavoriteListings.name] });
-      toast.success("Обявата е премахната от запазени обяви");
+      toast.success(t("The_listing_has_been_removed_from_favorites"));
     },
     onError: (error) => {
       handleError(error ?? null);