+++ app/context/TRQs/listings/mutations/useCreateListing.ts
@@ -6,8 +6,10 @@ import { createListingAsync } from "pawpal-fe-common/listings-api";
 import { ICreateListingPayload } from "pawpal-fe-common/listings-interfaces";
 import toast from "react-hot-toast";
 import useListings from "../useListings";
+import { useTranslation } from "react-i18next";
 
 const useCreateListing = () => {
+  const { t } = useTranslation();
   const queryClient = useQueryClient();
   const { handleError } = useGlobalErrorHandler();
 
@@ -16,7 +18,7 @@ const useCreateListing = () => {
       createListingAsync(payload),
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: [useListings.name] });
-      toast.success("Успешно създадохте обява");
+      toast.success(t("You_are_now_a_pet_sitter"));
     },
     onError: (error) => {
       handleError(error ?? null);