+++ app/context/TRQs/users/mutations/useAuthenticate.ts
@@ -7,14 +7,11 @@ import { IAuthenticatePayload } from "pawpal-fe-common/users-interfaces";
 import useCurrentUser from "../useCurrentUser";
 import { useAuth } from "@/app/context/AuthContext";
 import useListings from "../../listings/useListings";
-import { useRouter } from "next/navigation";
-import toast from "react-hot-toast";
 
 const useAuthenticate = (onSuccessCallback?: () => void) => {
   const queryClient = useQueryClient();
   const { handleError } = useGlobalErrorHandler();
   const { setAuthStatus } = useAuth();
-  const router = useRouter();
 
   return useMutation({
     mutationFn: async (payload: IAuthenticatePayload) =>