+++ app/layout.tsx
@@ -5,7 +5,6 @@ import ToasterProvider from "./providers/ToasterProvider";
 import ClientOnly from "@/app/components/ClientOnly";
 import { Analytics } from "@vercel/analytics/react";
 import { SpeedInsights } from "@vercel/speed-insights/next";
-import { ReduxProvider } from "./providers/ReduxProvider";
 import TanStackReactQueryProvider from "./providers/TanStackReactQueryProvider";
 import { AuthProvider } from "./context/AuthContext";
 import { GoogleOAuthProvider } from "@react-oauth/google";
@@ -27,14 +26,12 @@ export default async function RootLayout({
       <body className={inter.className}>
         <AuthProvider>
           <TanStackReactQueryProvider>
-            <ReduxProvider>
-              <GoogleOAuthProvider clientId="508483481007-bnvs2k6jpc0ei0t075ecps22b5cbga8f.apps.googleusercontent.com">
-                <ClientOnly>
-                  <ToasterProvider />
-                  <div>{children}</div>
-                </ClientOnly>
-              </GoogleOAuthProvider>
-            </ReduxProvider>
+            <GoogleOAuthProvider clientId="508483481007-bnvs2k6jpc0ei0t075ecps22b5cbga8f.apps.googleusercontent.com">
+              <ClientOnly>
+                <ToasterProvider />
+                <div>{children}</div>
+              </ClientOnly>
+            </GoogleOAuthProvider>
           </TanStackReactQueryProvider>
         </AuthProvider>
         <Analytics />